Start making a new main class so that the clocker class only has actual ROE information
