+++ arctic/__init__.py
@@ -1,3 +1,4 @@
+from arctic.main import ArcticMain
 from arctic.clock import Clocker
 from arctic.ccd_volume import CCDVolume, CCDVolumeComplex
 from arctic.traps import (
